Move the ecalls to a new crate
To resolve the (false-positive) dependency cycle we're moving the ecalls
to a separate crate which can be depended on from `getrandom`.
