+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = ["app-sdk", "client-sdk", "macros", "common", "bench", "cargo-vnd"]
+members = ["app-sdk", "client-sdk", "ecalls", "macros", "common", "bench", "cargo-vnd", "ecalls"]
