+++ ecalls/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "vanadium-ecalls"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+common = { path = "../common" }
